third_party/blink/web_tests/platform/generic/fast/events/drag-dataTransferItemList-file-handling-expected.txt

To manually test, select a text file for the file input element, and then drag 'Drag Me' to 'Drop Here' and drop. 

Drag Me
Drop Here

Tests DataTransferItemList file handling

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testItem.kind is "file"
PASS testItem.type is "text/plain"
PASS testItems.length is 1
PASS testItems.add(testFile) is null
PASS testItems.length is 1
PASS testItems[0].kind is "file"
PASS testItems[0].type is "text/plain"
PASS successfullyParsed is true

TEST COMPLETE
